backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
bcrypt==4.1.1
pyjwt==2.8.0
python-multipart==0.0.6
pydantic[email]==2.5.0
aiofiles==23.2.1
python-dateutil==2.8.2
email-validator==2.1.0
